Flat structure.
