+++ examples/Examples/Common/Example.cs
