+++ examples/Examples/Common/ExampleOptions.cs
