+++ examples/Examples/Common/ExampleOptionsWithKeys.cs
@@ -3,15 +3,18 @@ using CommandLine;
 namespace Concordium.Sdk.Examples;
 
 /// <summary>
-/// Command line options for the runnable SDK examples
-/// relating to account transactions.
+/// Command line options for the runnable SDK examples.
 ///
-/// This class can be used to specify the path to a file
+/// Like <see cref="ExampleOptions"/> with a further
+/// option for specifying a path the path to a file
 /// containing a supported import formats of
 /// <see cref="Concordium.Sdk.Wallets.WalletAccount"/>
 /// at the command-line.
+///
+/// This is useful when writing examples that work with
+/// account transactions.
 /// </summary>
-public abstract class TransactionExampleOptions : ExampleOptions
+public abstract class ExampleOptionsWithKeys : ExampleOptions
 {
     [Option(
         'k',
