+++ examples/Examples/RawClient/GetBlockTransactionEvents/Program.cs
@@ -9,7 +9,7 @@ namespace Concordium.Sdk.Examples.RawClient;
 /// <see cref="RawClient"/> wraps methods of the Concordium Node GRPC API V2 that were generated
 /// from the protocol buffer schema by the <see cref="Grpc.Core"/> library.
 /// </summary>
-class GetBlockTransactionEventsExample
+class Program
 {
     async static void GetBlockTransactionEvents(GetBlockTransactionEventsOptions options)
     {
