+++ examples/Examples/Transactions/RegisterData/Program.cs
@@ -6,18 +6,18 @@ using Concordium.Sdk.Client;
 namespace Concordium.Sdk.Examples.Transactions;
 
 /// <summary>
-/// Example demonstrating how to submit a transfer transaction.
+/// Example demonstrating how to submit a register data transaction.
 ///
 /// The example assumes you have your account key information stored
 /// in the Concordium browser wallet key export format, and that a path
 /// pointing to it is supplied to it from the command line.
 ///
-/// Cfr. <see cref="TransferTransactionExampleOptions"/> for more info
-/// on how to run the program, or refer to the help message.
+/// Cfr. <see cref="RegisterDataTransactionExampleOptions"/> for more
+/// info on how to run the program, or refer to the help message.
 /// </summary>
-class TransferTransactionExample
+class Program
 {
-    static void SendTransferTransaction(TransferTransactionExampleOptions options)
+    static void SendRegisterDataTransaction(RegisterDataTransactionExampleOptions options)
     {
         // Read the account keys from a file.
         string walletData = File.ReadAllText(options.WalletKeysFile);
@@ -30,35 +30,35 @@ class TransferTransactionExample
             60 // Use a timeout of 60 seconds.
         );
 
-        // Create the transfer transaction.
-        CcdAmount amount = CcdAmount.FromCcd(options.Amount);
-        AccountAddress receiver = AccountAddress.From(options.Receiver);
-        Transfer transferPayload = new Transfer(amount, receiver);
+        // Create the register data transaction.
+        // Encode a singleton string as CBOR and use that as the data to register.
+        OnChainData data = OnChainData.FromTextEncodeAsCBOR(options.Data);
+        RegisterData transferPayload = new RegisterData(data);
 
         // Prepare the transaction for signing.
         AccountAddress sender = account.AccountAddress;
         AccountSequenceNumber nonce = client.GetNextAccountSequenceNumber(sender);
         Expiry expiry = Expiry.AtMinutesFromNow(30);
-        PreparedAccountTransaction<Transfer> preparedTransfer = transferPayload.Prepare(
+        PreparedAccountTransaction<RegisterData> preparedTransfer = transferPayload.Prepare(
             sender,
             nonce,
             expiry
         );
 
         // Sign the transaction using the account keys.
-        SignedAccountTransaction<Transfer> signedTransfer = preparedTransfer.Sign(account);
+        SignedAccountTransaction<RegisterData> signedTransfer = preparedTransfer.Sign(account);
 
-        // Submit the transaction.
+        // Try to submit the transaction.
         TransactionHash txHash = client.SendTransaction(signedTransfer);
 
         // Print the transaction hash.
         Console.WriteLine(
-            $"Successfully submitted transfer transaction with hash {txHash.ToString()}"
+            $"Successfully submitted register data transaction with hash {txHash.ToString()}"
         );
     }
 
     static void Main(string[] args)
     {
-        Example.Run<TransferTransactionExampleOptions>(args, SendTransferTransaction);
+        Example.Run<RegisterDataTransactionExampleOptions>(args, SendRegisterDataTransaction);
     }
 }
