+++ examples/Examples/Transactions/Transfer/Concordium.Sdk.Examples.Transactions.Transfer.csproj
@@ -8,7 +8,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Concordium.Sdk.Examples.csproj" />
+    <ProjectReference Include="..\..\Common\Concordium.Sdk.Examples.Common.csproj" />
     <ProjectReference Include="..\..\..\..\src\Concordium.Sdk.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="CommandLineParser" Version="2.9.1" />
   </ItemGroup>
 </Project>
